+++ CHANGELOG.md
@@ -15,6 +15,7 @@ omnibus-gitlab repository.
 - Enable support for git push options in the git config (Romain Maffina) 08edd3e4
 - Update Redis to 3.2.11 (Takuya Noguchi)
 - Turn on postgresql SSL by default b18597e3
+- [GitLab Geo] Add a command to promote secondary node to primary
 
 10.2.3